cleaning up week selection
