+++ frontend/cfb/resources/game.js
@@ -185,7 +185,7 @@ async function processPlays(gameId) {
 }
 
 
-async function retrieveGameList(url, params) {
+async function getGames(url, params) {
     var gameList = await getSchedule(params);
     if (gameList == null) {
         throw Error(`Data not available for ${url} because of a service error.`)
@@ -233,7 +233,33 @@ async function retrieveGameList(url, params) {
     return gameList;
 }
 
+async function routeGameList(req, res, next, payload) {
+    try {
+        let gameList = await getGames(req.originalUrl, payload);
+        let weekList = Schedule.getWeeksMap();
+        let groupList = Schedule.getGroups();
+
+        var weekTitle = null;
+        if (payload["year"]) {
+            weekTitle = weekList[payload["year"]]?.find(w => parseInt(w.type) == parseInt(payload["type"]) && parseInt(w.value) == parseInt(payload["week"]))?.title;
+        }
+
+        return res.render('pages/cfb/index', {
+            scoreboard: gameList,
+            weekList: weekList,
+            groups: groupList,
+            year: payload["year"],
+            week: payload["week"],
+            seasontype: payload["type"],
+            group: payload["group"] || 80,
+            title: weekTitle
+        });
+    } catch(err) {
+        return next(err)
+    }
+}
+
 module.exports = {
-    retrieveGameList,
+    routeGameList,
     retrievePBP
 }
